ailand_AI_assistant/requirements.txt
fastapi
uvicorn
pydantic
SPARQLWrapper
python-dotenv
requests
pydantic-settings
google-generativeai
